[X86] Extract HiPE prologue constants into metadata
X86FrameLowering::adjustForHiPEPrologue() contains a hard-coded offset
into an Erlang Runtime System-internal data structure (the PCB). As the
layout of this data structure is prone to change, this poses problems
for maintaining compatibility.

To address this problem, we will have the compiler spit out this
information as module-level named metadata, for example (where
P_NSP_LIMIT is the offending offset):

    ...
    !hipe.literals = !{ !2, !3, !4 }
    !2 = !{ !"P_NSP_LIMIT", i32 152 }
    !3 = !{ !"X86_LEAF_WORDS", i32 24 }
    !4 = !{ !"AMD64_LEAF_WORDS", i32 24 }
    ...