[X86] Fix tailcall return address clobber bug
This bug was introduced by r237977, which refactored the tail call
sequence to be generated in two passes instead of one.

Unfortunately, the stack adjustment produced by the first pass was not
recognized by X86FrameLowering::mergeSPUpdates() in all cases, causing
code such as the following, which clobbers the return address, to be
generated:

	popl	%edi
	popl	%edi
	pushl	%eax
	jmp	tailcallee              # TAILCALL

By generalizing X86FrameLowering::mergeSPUpdates() to recognize a pop
instruction as a stack adjustment when the register is not used by the
following return instruction, this bug is fixed.